Restructured to take a list of changes so that we can use this for Major editing

+++ algorithms/TestReadCharLengthAndModifyInFile/models/SetPositionWithTextInfoModel.cs
@@ -3,6 +3,6 @@ namespace TestReadCharLengthAndModifyInFile.models;
 public class SetPositionWithTextInfoModel
 {
     public long StartPosition { get; set; }
-    public long Length { get; set; }
+    public int Length { get; set; }
     public string Text { get; set; }
 }
